+++ src/lib.rs
@@ -2,6 +2,6 @@
 extern crate nom;
 
 pub mod color_parser;
-pub mod operation;
+pub mod expression;
 pub mod operators;
 pub mod traits;
